+++ src/test/java/org/matsim/ptFares/PtFaresHandlerTest.java
@@ -1,9 +1,7 @@
 package org.matsim.ptFares;
 
-import ch.sbb.matsim.config.SBBTransitConfigGroup;
+
 import ch.sbb.matsim.config.SwissRailRaptorConfigGroup;
-import ch.sbb.matsim.mobsim.qsim.SBBTransitModule;
-import ch.sbb.matsim.mobsim.qsim.pt.SBBTransitEngineQSimModule;
 import ch.sbb.matsim.routing.pt.raptor.RaptorIntermodalAccessEgress;
 import ch.sbb.matsim.routing.pt.raptor.SwissRailRaptorModule;
 import org.apache.log4j.Logger;
@@ -27,8 +25,6 @@ import org.matsim.core.controler.Controler;
 import org.matsim.core.controler.OutputDirectoryHierarchy;
 import org.matsim.core.gbl.Gbl;
 import org.matsim.core.scenario.ScenarioUtils;
-import org.matsim.parkingCost.ParkingCostConfigGroup;
-import org.matsim.parkingCost.ParkingCostModule;
 import org.matsim.pt.transitSchedule.api.TransitSchedule;
 import org.matsim.pt.transitSchedule.api.TransitStopFacility;
 
@@ -37,10 +33,12 @@ import java.util.*;
 import static org.matsim.core.config.groups.ControlerConfigGroup.RoutingAlgorithmType.FastAStarLandmarks;
 
 public class PtFaresHandlerTest {
-    private static final Logger log = Logger.getLogger(PtFaresHandlerTest.class );
+    private static final Logger log = Logger.getLogger(PtFaresHandlerTest.class);
+
+    Map<Id<Person>, Double> transitFare = new HashMap<>();
 
     @Test
-    public final void testPtFaresHandlerTest(){
+    public final void testPtFaresHandlerTest_v2() {
 
         String configPath = "./test/input/config.xml";
 
@@ -48,47 +46,41 @@ public class PtFaresHandlerTest {
         Config config = prepareConfig(configPath);
 
         // Prepare scenario
-        Scenario scenario = prepareScenario(config) ;
-        Controler controler = prepareControler(scenario) ;
+        Scenario scenario = prepareScenario(config);
+        Controler controler = prepareControler(scenario);
 
         EventsManager events = controler.getEvents();
-        Tester tester = new Tester();
+        PtFaresHandlerTest.EventsTester tester = new PtFaresHandlerTest.EventsTester();
         events.addHandler(tester);
-
 
         controler.run();
 
-        tester.getPerson2Purpose().entrySet().stream().forEach(entry ->{
-            System.out.printf("Person %$1s enters vehicle %$2d", entry.getKey().toString(), entry.getValue());
-        });
-
-        //
-
         // Check output param
-        Assert.equals(scenario.getPopulation().getPersons().get(Id.createPersonId("1")).getSelectedPlan().getScore(), 200);
-        Assert.equals(scenario.getPopulation().getPersons().get(Id.createPersonId("2")).getSelectedPlan().getScore(), 200);
-        Assert.equals(scenario.getPopulation().getPersons().get(Id.createPersonId("3")).getSelectedPlan().getScore(), 100);
-        Assert.equals(scenario.getPopulation().getPersons().get(Id.createPersonId("4")).getSelectedPlan().getScore(), 100);
-        Assert.equals(scenario.getPopulation().getPersons().get(Id.createPersonId("5")).getSelectedPlan().getScore(), 100);
-        Assert.equals(scenario.getPopulation().getPersons().get(Id.createPersonId("6")).getSelectedPlan().getScore(), 200);
+        Assert.equals(transitFare.get(Id.createPersonId("1")), -300.);
+        Assert.equals(transitFare.get(Id.createPersonId("2")), -200.);
+        Assert.equals(transitFare.get(Id.createPersonId("3")), -100.);
+        Assert.equals(transitFare.get(Id.createPersonId("4")), -100.);
+        Assert.equals(transitFare.get(Id.createPersonId("5")), -100.);
+        Assert.equals(transitFare.get(Id.createPersonId("6")), -200.);
+        Assert.isTrue(! transitFare.containsKey(Id.createPersonId("7")));
 
     }
 
 
-    public static Controler prepareControler( Scenario scenario ) {
+    public static Controler prepareControler(Scenario scenario) {
         Gbl.assertNotNull(scenario);
 
-        final Controler controler = new Controler( scenario );
+        final Controler controler = new Controler(scenario);
 
         // -- ADDITIONAL MODULES --
         if (controler.getConfig().transit().isUsingTransitInMobsim()) {
             // use the sbb pt raptor router
-            controler.addOverridingModule( new AbstractModule() {
+            controler.addOverridingModule(new AbstractModule() {
                 @Override
                 public void install() {
-                    install( new SwissRailRaptorModule() );
+                    install(new SwissRailRaptorModule());
                 }
-            } );
+            });
         } else {
             log.warn("Public transit will be teleported and not simulated in the mobsim! "
                     + "This will have a significant effect on pt-related parameters (travel times, modal split, and so on). "
@@ -96,65 +88,49 @@ public class PtFaresHandlerTest {
         }
 
         // use the (congested) car travel time for the teleported ride modes
-        controler.addOverridingModule( new AbstractModule() {
+        controler.addOverridingModule(new AbstractModule() {
             @Override
             public void install() {
-                addTravelTimeBinding( TransportMode.ride ).to( networkTravelTime() );
-                addTravelDisutilityFactoryBinding( TransportMode.ride ).to( carTravelDisutilityFactoryKey() ); }
-        } );
+                addTravelTimeBinding(TransportMode.ride).to(networkTravelTime());
+                addTravelDisutilityFactoryBinding(TransportMode.ride).to(carTravelDisutilityFactoryKey());
+            }
+        });
 
         // use scoring parameters for intermodal PT routing
-        controler.addOverridingModule( new AbstractModule() {
+        controler.addOverridingModule(new AbstractModule() {
             @Override
             public void install() {
                 bind(RaptorIntermodalAccessEgress.class).to(org.matsim.run.StuttgartRaptorIntermodalAccessEgress.class);
             }
-        } );
-
-/*        // use deterministic transport simulation of SBB
-        controler.addOverridingModule(new AbstractModule() {
-            @Override
-            public void install() {
-                // To use the deterministic pt simulation (Part 1 of 2):
-                install(new SBBTransitModule());
-            }
-
-            // To use the deterministic pt simulation (Part 2 of 2):
         });
-
-        controler.configureQSimComponents(components -> {
-            SBBTransitEngineQSimModule.configure(components);
-        });*/
-
-        // use parking cost module
-        controler.addOverridingModule(new ParkingCostModule());
 
         // use pt fares module
         controler.addOverridingModule(new PtFaresModule());
 
         return controler;
+
     }
 
 
     private Scenario prepareScenario(Config config) {
 
-        Gbl.assertNotNull( config );
+        Gbl.assertNotNull(config);
 
         Scenario scenario = ScenarioUtils.loadScenario(config);
 
         // write fareZones into transitScheduleFile
         TransitSchedule schedule = scenario.getTransitSchedule();
 
-        schedule.getFacilities().get(Id.create("1", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "2");
-        schedule.getFacilities().get(Id.create("2a", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "2");
-        schedule.getFacilities().get(Id.create("2b", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "2");
-        schedule.getFacilities().get(Id.create("3", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "1,2");
+        schedule.getFacilities().get(Id.create("1", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "1");
+        schedule.getFacilities().get(Id.create("2a", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "2");
+        schedule.getFacilities().get(Id.create("2b", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "2");
+        schedule.getFacilities().get(Id.create("3", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "1,2");
 
 
-        schedule.getFacilities().get(Id.create("4", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "2");
-        schedule.getFacilities().get(Id.create("5a", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "2");
-        schedule.getFacilities().get(Id.create("5b", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "2");
-        schedule.getFacilities().get(Id.create("6", TransitStopFacility.class)).getAttributes().putAttribute("FareZone", "2");
+        schedule.getFacilities().get(Id.create("4", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "2");
+        schedule.getFacilities().get(Id.create("5a", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "2");
+        schedule.getFacilities().get(Id.create("5b", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "2");
+        schedule.getFacilities().get(Id.create("6", TransitStopFacility.class)).getAttributes().putAttribute("ptFareZone", "out");
 
         return scenario;
     }
@@ -162,8 +138,8 @@ public class PtFaresHandlerTest {
 
     private Config prepareConfig(String configPath) {
 
-        ConfigGroup[] customModulesToAdd = new ConfigGroup[]{ setupRaptorConfigGroup(), setupPTFaresGroup(), new ParkingCostConfigGroup()};
-        // ConfigGroup[] customModulesToAdd = new ConfigGroup[]{ setupRaptorConfigGroup(), setupPTFaresGroup(), setupSBBTransit(), new ParkingCostConfigGroup()};
+        // Add custom modules
+        ConfigGroup[] customModulesToAdd = new ConfigGroup[]{setupPTFaresGroup(), setupRaptorConfigGroup()};
         ConfigGroup[] customModulesAll = new ConfigGroup[customModulesToAdd.length];
 
         int counter = 0;
@@ -174,7 +150,7 @@ public class PtFaresHandlerTest {
         }
 
         // -- LOAD CONFIG WITH CUSTOM MODULES
-        final Config config = ConfigUtils.loadConfig( configPath, customModulesAll );
+        final Config config = ConfigUtils.loadConfig(configPath, customModulesAll);
 
 
         // -- CONTROLER --
@@ -183,29 +159,20 @@ public class PtFaresHandlerTest {
 
 
         // -- VSP DEFAULTS --
-        config.vspExperimental().setVspDefaultsCheckingLevel( VspExperimentalConfigGroup.VspDefaultsCheckingLevel.ignore );
+        config.vspExperimental().setVspDefaultsCheckingLevel(VspExperimentalConfigGroup.VspDefaultsCheckingLevel.ignore);
         config.plansCalcRoute().setAccessEgressType(PlansCalcRouteConfigGroup.AccessEgressType.accessEgressModeToLink); //setInsertingAccessEgressWalk( true );
-        config.qsim().setUsingTravelTimeCheckInTeleportation( true );
-        config.qsim().setTrafficDynamics( QSimConfigGroup.TrafficDynamics.kinematicWaves );
+        config.qsim().setUsingTravelTimeCheckInTeleportation(true);
+        config.qsim().setTrafficDynamics(QSimConfigGroup.TrafficDynamics.kinematicWaves);
 
 
         // -- OTHER --
         config.qsim().setUsePersonIdForMissingVehicleId(false);
         config.controler().setRoutingAlgorithmType(FastAStarLandmarks);
-        config.subtourModeChoice().setProbaForRandomSingleTripMode( 0.5 );
-        config.qsim().setInsertingWaitingVehiclesBeforeDrivingVehicles( true );
+        config.subtourModeChoice().setProbaForRandomSingleTripMode(0.5);
+        config.qsim().setInsertingWaitingVehiclesBeforeDrivingVehicles(true);
 
-        return config ;
+        return config;
 
-    }
-
-
-    private static SwissRailRaptorConfigGroup setupRaptorConfigGroup() {
-
-        SwissRailRaptorConfigGroup configRaptor = new SwissRailRaptorConfigGroup();
-        configRaptor.setUseIntermodalAccessEgress(false);
-
-        return configRaptor;
     }
 
 
@@ -221,63 +188,26 @@ public class PtFaresHandlerTest {
         configFares.addZonePriceSettings(paramSetZone1);
 
         PtFaresConfigGroup.ZonePricesParameterSet paramSetZone2 = new PtFaresConfigGroup.ZonePricesParameterSet();
-        paramSetZone1.setNumberZones(2);
-        paramSetZone1.setTicketPrice(200.);
+        paramSetZone2.setNumberZones(2);
+        paramSetZone2.setTicketPrice(200.);
         configFares.addZonePriceSettings(paramSetZone2);
 
         return configFares;
     }
 
 
-    private static SBBTransitConfigGroup setupSBBTransit() {
+    private static SwissRailRaptorConfigGroup setupRaptorConfigGroup() {
+        SwissRailRaptorConfigGroup configRaptor = new SwissRailRaptorConfigGroup();
+        configRaptor.setUseIntermodalAccessEgress(false);
 
-        SBBTransitConfigGroup sbbTransit = new SBBTransitConfigGroup();
-        Set<String> modes = new HashSet<>(Arrays.asList(new String[]{"train"}));
-        sbbTransit.setDeterministicServiceModes(modes);
-        sbbTransit.setCreateLinkEventsInterval(1);
-
-        return sbbTransit;
+        return configRaptor;
     }
 
 
-    private static class FareSumCalculator implements PersonMoneyEventHandler {
-        Map<Id<Person>, Double> person2Fare = new HashMap<>();
-
+    private final class EventsTester implements PersonMoneyEventHandler {
         @Override
         public void handleEvent(PersonMoneyEvent event) {
-
-            person2Fare.put(event.getPersonId(), event.getAmount());
-
-        }
-
-        @Override
-        public void reset(int iteration) {
-            person2Fare.clear();
-        }
-
-        private Map<Id<Person>, Double> getPerson2Fare() {
-            return person2Fare;
-        }
-    }
-
-
-    private static class Tester implements PersonMoneyEventHandler {
-        Map<Id<Person>, String> person2Purpose = new HashMap<>();
-
-        @Override
-        public void handleEvent(PersonMoneyEvent event) {
-
-            person2Purpose.put(event.getPersonId(), event.getPurpose());
-
-        }
-
-        @Override
-        public void reset(int iteration) {
-            person2Purpose.clear();
-        }
-
-        private Map<Id<Person>, String> getPerson2Purpose() {
-            return person2Purpose;
+            transitFare.put(event.getPersonId(), event.getAmount());
         }
     }
