+++ test/input/config.xml
@@ -176,7 +176,7 @@
 
     <module name="transitRouter">
         <!-- Factor with which direct walk generalized cost is multiplied before it is compared to the pt generalized cost.  Set to a very high value to reduce direct walk results. -->
-        <param name="directWalkFactor" value="1.0"/>
+        <param name="directWalkFactor" value="10.0"/>
         <!-- maximum beeline distance between stops that agents could transfer to by walking -->
         <param name="maxBeelineWalkConnectionDistance" value="300.0"/>
     </module>
