+++ test/input/plans.xml
@@ -43,13 +43,13 @@
 		</plan>
 
 	</person>
-	
+
 	<person id="4" employed="no">
 		<plan selected="yes">
 			<act type="h"  x="1000.0" y="3000.0" end_time="06:18:16" />
 			<leg mode="pt" >				
 			</leg>
-			<act type="w"  x="3980.0" y="3080.0" end_time="14:43:36" />
+			<act type="w"  x="2040.0" y="2800.0" end_time="14:43:36" />
 			<leg mode="pt">			
 			</leg>
 			<act type="h"  x="1000.0" y="3000.0" end_time="17:43:36"/>
@@ -60,15 +60,14 @@
 			</leg>
 			<act type="h"  x="1000.0" y="3000.0" />
 		</plan>
-
 	</person>
-	
+
 	<person id="5" employed="no">
 		<plan selected="yes">
 			<act type="h"  x="1000.0" y="3000.0" end_time="07:18:16" />
 			<leg mode="pt" >				
 			</leg>
-			<act type="w"  x="3980.0" y="3080.0" end_time="15:43:36" />
+			<act type="w"  x="2040.0" y="2800.0" end_time="15:43:36" />
 			<leg mode="pt">				
 			</leg>
 			<act type="h"  x="1000.0" y="3000.0" end_time="17:43:36" />
@@ -87,7 +86,7 @@
 			<act type="h"  x="1000.0" y="3000.0" end_time="07:18:16" />
 			<leg mode="pt" >				
 			</leg>
-			<act type="w"  x="3980.0" y="3080.0" end_time="17:43:36" />
+			<act type="w"  x="2040.0" y="2800.0" end_time="16:43:36" />
 			<leg mode="pt">				
 			</leg>
 			<act type="h"  x="1000.0" y="3000.0" end_time="18:43:36" />
@@ -97,6 +96,19 @@
 			<leg mode="pt">				
 			</leg>
 			<act type="h"  x="1000.0" y="3000.0" />
+		</plan>
+
+	</person>
+	
+	<person id="7" employed="no">
+		<plan selected="yes">
+			<act type="h"  x="1040.0" y="1040.0" end_time="07:18:16" />
+			<leg mode="car" >				
+			</leg>
+			<act type="w"  x="3960.0" y="1100.0" end_time="17:43:36" />
+			<leg mode="car">				
+			</leg>
+			<act type="h"  x="1040.0" y="1040.0" />
 		</plan>
 
 	</person>
